FIXME: ASoC: q6dsp: q6afe: Tack on client_handle tracking
Each vote for the clock has a handle value that has to be used to unvote
the vote. THe vote handle was not tracked so each unvote failed and the
vote tracker on adsp probably just increased forever. this breaks sleep.

Add client_handle tracking to make sure votes are unvoted and the soc
can sleep when the hw is unused.

also set token to 0 on devote since this is what downstream does.

Fixes: 55e07531d922 ("ASoC: q6dsp: q6afe: add lpass hw voting support")
Signed-off-by: Nikita Travkin <nikita@trvn.ru>