soc: qcom: smp2p: Drop redundant stack copies of entry names
For the purposes of this driver, a char is always going to be the same size
as an u8, so we can access the entry names directly instead of making a
copy on the stack. Several other Qualcomm-related drivers use char directly
in such binary structs as well (e.g. qcom_battmgr).

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>